pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "wikipedia-client"
description = "Wikpedia API client and simple wikitext parser"
authors = [
  { name="Wojciech Pietrzok" },
]
license = {file = "LICENSE"}
# dynamic = [ "version", "readme", "dependencies" ]
dynamic = [ "version", "readme" ]
dependencies = [
  "requests==2.28.1",
  "python-dateutil==2.8.2",
  "sql-query-builder @ git+https://github.com/kosciak/sql-query-builder"
]
requires-python = ">=3.8"

[project.urls]
"Homepage" = "https://github.com/kosciak/wikipedia-client"

[tool.setuptools]
packages = ["wikipedia"]

[tool.setuptools.dynamic]
version = {attr = "wikipedia.__version__"}
readme = {file = ["README.md"]}
# dependencies = {file = ["requirements.txt"]}

